Made a brand new beta build scene
This time I used the Rigidbody FPS controller and All the levels are working as inteded!
Edited a lot of scripts for better performance so we weren't constantly calling the update method and just a ton of changes. probably should ahve broken this up into a few commits. oh well.